fix bonus no float issue

+++ Assets/Scripts/Character.cs
@@ -24,9 +24,9 @@ public class Character : MonoBehaviour
 
     void Start()
     {
-        damage += damage * GameManager.Instance.abilities.FindAll(x => x == ItemType.Claws).Count / 10;
+        damage += damage * ((float) GameManager.Instance.abilities.FindAll(x => x == ItemType.Claws).Count) / 10;
 
-        int numFogs = GameManager.Instance.abilities.FindAll(x => x == ItemType.Fog).Count;
+        float numFogs = GameManager.Instance.abilities.FindAll(x => x == ItemType.Fog).Count;
         Vector3 localScale = fogOfWar.transform.localScale;
         localScale.x += numFogs * 0.15f;
         localScale.y += numFogs * 0.15f;
